flatpak: Autospec creation for update from version 1.0.4 to version 1.0.5
Alexander Larsson (6):
      run: Make repos optional (again)
      Don't allow writes to runtime files in /etc
      Make per-app generated files read-only
      Update NEWS for release
      Bump version to 1.0.5
      Update pofiles

Matthias Clasen (3):
      remote: Be robust against bad names
      transaction: Don't call g_object_ref on a GBytes
      transaction: Improve the docs

Piotr Drąg (2):
      Update Polish translation
      Update Polish translation

Changes in 1.0.5
================

There was a sandbox bug in the previous version where parts of the runtime
/etc was not mounted read-only. In case the runtime was installed as the
user (not the default) this means that the app could modify files on the
runtime. Nothing in the host uses the runtime files, so this is not a direct
sandbox escape, but it is possible that an app can confuse a different app
that has higher permissions and so gain privileges.

So, it is recommended that everyone shipping flatpak to update to
1.0.5, or at least backport the change in commit
6711d7ae99c50a9dca8e4e2e9e9989a8fa6c3f06.

Changes in this version:

(NEWS truncated at 15 lines)