flatpak: Autospec creation for update from version 1.0.6 to version 1.0.7
Alexander Larsson (4):
      Don't expose /proc when running apply_extra
      Bump version to 1.0.7
      Update NEWS for release
      Update pofiles

AsciiWolf (1):
      Update Czech translation

Changes in 1.0.7
================

The CVE-2019-5736 runc vulnerability is about using /proc/self/exe
to modify the host side binary from the sandbox. This mostly does not
affect flatpak since the flatpak sandbox is not run with root permissions.
However, there is one case (running the apply_extra script for system
installs) where this happens, so this release contains a fix for that.

 * Don't expose /proc in apply_extra script sandbox.

CVEs fixed in this build:
CVE-2019-5736